Move sample command to opuscleaner-datasets
