+++ pyproject.toml
@@ -9,23 +9,17 @@ readme = "README.md"
 requires-python = ">=3.9"
 license = "MIT"
 keywords = []
-authors = [
-  { name = "Jelmer van der Linde", email = "jelmer@ikhoefgeen.nl" },
-]
+authors = [{ name = "Jelmer van der Linde", email = "jelmer@ikhoefgeen.nl" }]
 classifiers = [
-  "Development Status :: 4 - Beta",
-  "Programming Language :: Python",
-  "Programming Language :: Python :: 3.9",
-  "Programming Language :: Python :: 3.10",
-  "Programming Language :: Python :: 3.11",
-  "Programming Language :: Python :: Implementation :: CPython",
-  "Programming Language :: Python :: Implementation :: PyPy",
+    "Development Status :: 4 - Beta",
+    "Programming Language :: Python",
+    "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: Implementation :: CPython",
+    "Programming Language :: Python :: Implementation :: PyPy",
 ]
-dynamic = [
-  "version",
-  "dependencies",
-  "optional-dependencies"
-]
+dynamic = ["version", "dependencies", "optional-dependencies"]
 
 [project.scripts]
 opuscleaner-server = "opuscleaner.server:main"
@@ -33,6 +27,7 @@ opuscleaner-clean = "opuscleaner.clean:main"
 opuscleaner-col = "opuscleaner.col:main"
 opuscleaner-threshold = "opuscleaner.threshold:main"
 opuscleaner-sample = "opuscleaner.sample:main"
+opuscleaner-datasets = "opuscleaner.datasets:main"
 opuscleaner-download = "opuscleaner.download:main"
 
 [project.urls]
@@ -44,10 +39,7 @@ Source = "https://github.com/hplt-project/opuscleaner"
 path = "opuscleaner/__about__.py"
 
 [tool.hatch.envs.default]
-dependencies = [
-  "pytest",
-  "pytest-cov",
-]
+dependencies = ["pytest", "pytest-cov"]
 [tool.hatch.envs.default.scripts]
 cov = "pytest --cov-report=term-missing --cov-config=pyproject.toml --cov=opuscleaner --cov=tests {args}"
 no-cov = "cov --no-cov {args}"
@@ -56,9 +48,7 @@ no-cov = "cov --no-cov {args}"
 python = ["39", "310", "311"]
 
 [tool.hatch.build]
-include = [
-  "/opuscleaner",
-]
+include = ["/opuscleaner"]
 
 [tool.hatch.build.force-include]
 "frontend/dist" = "/opuscleaner/frontend"
@@ -72,13 +62,7 @@ all = ["requirements-all.txt"]
 [tool.coverage.run]
 branch = true
 parallel = true
-omit = [
-  "opuscleaner/__about__.py",
-]
+omit = ["opuscleaner/__about__.py"]
 
 [tool.coverage.report]
-exclude_lines = [
-  "no cov",
-  "if __name__ == .__main__.:",
-  "if TYPE_CHECKING:",
-]
+exclude_lines = ["no cov", "if __name__ == .__main__.:", "if TYPE_CHECKING:"]
